Ability to handle directories as input
When passed a folder as input, the program will attempt to recursively
create a folder with the name specified in the `-o` option. Or if there
is already a folder with the same name as the output location, the
program will write each of the input files to that folder. If the input
comes from stdin and the output location is a folder, a file named
`htmlmin.html` will be saved in that folder.